requirements.txt
asgiref==3.9.1
certifi==2025.8.3
Django==5.2.5
django-email-users==2.1.0
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
phonenumberslite==9.0.11
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.5.0
